flask_api/requirements.txt
Flask
Jinja2
cohere
pinecone
openai
python-dotenv
PyPDF2
